requirements.txt
numpy
pandas
nibabel
tqdm
scikit-learn
torch
monai
PyYAML
openpyxl 